Fix error re-install supersu

+++ scripts/install_supersu.sh
@@ -14,9 +14,15 @@ rm system/bin/app_process
 ln -s /system/xbin/daemonsu system/bin/app_process
 
 if [ ! -z "$ROOT_SURVIVAL" ] || [ ! -z "$SLIM_DOWN" ]; then
-  mv system/bin/app_process32 system/bin/app_process32_original
+  if [ ! -f system/bin/app_process32_original ]; then
+    mv system/bin/app_process32 system/bin/app_process32_original
+  else 
+    rm system/bin/app_process32
+  fi
   ln -s /system/xbin/daemonsu system/bin/app_process32
-  cp system/bin/app_process32_original system/bin/app_process_init
+  if [ ! -f system/bin/app_process_init ]; then
+    cp system/bin/app_process32_original system/bin/app_process_init
+  fi
   
   cp system/etc/install-recovery.sh system/etc/install-recovery.sh.bak  
 fi
